+++ engine/src/dpx_bmpimage.rs
@@ -176,7 +176,7 @@ unsafe extern "C" fn get_density(
     mut ydensity: *mut f64,
     mut hdr: *mut hdr_info,
 ) {
-    if (*hdr).x_pix_per_meter > 0i32 as u32 && (*hdr).y_pix_per_meter > 0i32 as u32 {
+    if (*hdr).x_pix_per_meter > 0_u32 && (*hdr).y_pix_per_meter > 0_u32 {
         /* 0 for undefined. FIXME */
         *xdensity = 72.0f64 / ((*hdr).x_pix_per_meter as f64 * 0.0254f64);
         *ydensity = 72.0f64 / ((*hdr).y_pix_per_meter as f64 * 0.0254f64)
@@ -196,15 +196,15 @@ pub unsafe extern "C" fn bmp_get_bbox(
     let mut r: i32 = 0;
     let mut hdr: hdr_info = {
         let mut init = hdr_info {
-            offset: 0i32 as u32,
-            hsize: 0i32 as u32,
-            width: 0i32 as u32,
+            offset: 0_u32,
+            hsize: 0_u32,
+            width: 0_u32,
             height: 0i32,
             compression: 0i32,
             bit_count: 0i32 as u16,
             psize: 0i32,
-            x_pix_per_meter: 0i32 as u32,
-            y_pix_per_meter: 0i32 as u32,
+            x_pix_per_meter: 0_u32,
+            y_pix_per_meter: 0_u32,
         };
         init
     };
@@ -260,15 +260,15 @@ pub unsafe extern "C" fn bmp_include_image(
     };
     let mut hdr: hdr_info = {
         let mut init = hdr_info {
-            offset: 0i32 as u32,
-            hsize: 0i32 as u32,
-            width: 0i32 as u32,
+            offset: 0_u32,
+            hsize: 0_u32,
+            width: 0_u32,
             height: 0i32,
             compression: 0i32,
             bit_count: 0i32 as u16,
             psize: 0i32,
-            x_pix_per_meter: 0i32 as u32,
-            y_pix_per_meter: 0i32 as u32,
+            x_pix_per_meter: 0_u32,
+            y_pix_per_meter: 0_u32,
         };
         init
     };
@@ -306,7 +306,7 @@ pub unsafe extern "C" fn bmp_include_image(
         num_palette = hdr
             .offset
             .wrapping_sub(hdr.hsize)
-            .wrapping_sub(14i32 as u32)
+            .wrapping_sub(14_u32)
             .wrapping_div(hdr.psize as u32) as i32;
         info.bits_per_component = hdr.bit_count as i32;
         info.num_components = 1i32
@@ -529,8 +529,8 @@ unsafe extern "C" fn read_header(mut handle: rust_input_handle_t, mut hdr: *mut 
     if ttstub_input_read(
         handle,
         p as *mut i8,
-        (*hdr).hsize.wrapping_sub(4i32 as u32) as size_t,
-    ) != (*hdr).hsize.wrapping_sub(4i32 as u32) as i64
+        (*hdr).hsize.wrapping_sub(4_u32) as size_t,
+    ) != (*hdr).hsize.wrapping_sub(4_u32) as i64
     {
         dpx_warning(b"Could not read BMP file header...\x00" as *const u8 as *const i8);
         return -1i32;
@@ -541,8 +541,8 @@ unsafe extern "C" fn read_header(mut handle: rust_input_handle_t, mut hdr: *mut 
             p = p.offset(2);
             (*hdr).height = *p.offset(0) as i32 + ((*p.offset(1) as i32) << 8i32);
             p = p.offset(2);
-            (*hdr).x_pix_per_meter = 0i32 as u32;
-            (*hdr).y_pix_per_meter = 0i32 as u32;
+            (*hdr).x_pix_per_meter = 0_u32;
+            (*hdr).y_pix_per_meter = 0_u32;
             if *p.offset(0) as i32 + ((*p.offset(1) as i32) << 8i32) != 1i32 {
                 dpx_warning(
                     b"Unknown bcPlanes value in BMP COREHEADER.\x00" as *const u8 as *const i8,
